refactor: :recycle: authorization service
service belongs to the domain layer, not to the adapter layer
