+++ core/service/authorization.go
@@ -16,6 +16,8 @@ type AuthorizationService struct {
 }
 
 func NewAuthorizationService(usecases *usecase.AuthorizationUsecase) *AuthorizationService {
+	log.Println("iniciando serviços gRPC")
+
 	return &AuthorizationService{
 		usecases: usecases,
 	}
@@ -24,7 +26,7 @@ func NewAuthorizationService(usecases *usecase.AuthorizationUsecase) *Authorizat
 // Process the authorization request and return to the gRPC caller
 func (a *AuthorizationService) Process(ctx context.Context, in *pb.AuthorizationRequest) (*pb.AuthorizationResponse, error) {
 	log.Println("service.authorizations.process")
-	
+
 	auth := &dto.AuthorizationProcessDto{
 		Value:         in.Value,
 		ClientID:      in.ClientId,
